services/buyer_server/requirements.txt
Flask==3.0.0
grpcio>=1.60.0
grpcio-tools>=1.60.0
